chore: move user_info to auth crate
