+++ src/auth/src/lib.rs
@@ -1,0 +1,3 @@
+mod user_info;
+
+pub use user_info::UserInfo;
