+++ src/auth/src/user_info.rs
@@ -1,0 +1,26 @@
+const DEFAULT_USERNAME: &str = "greptime";
+
+#[derive(Clone, Debug)]
+pub struct UserInfo {
+    username: String,
+}
+
+impl Default for UserInfo {
+    fn default() -> Self {
+        Self {
+            username: DEFAULT_USERNAME.to_string(),
+        }
+    }
+}
+
+impl UserInfo {
+    pub fn username(&self) -> &str {
+        self.username.as_str()
+    }
+
+    pub fn new(username: impl Into<String>) -> Self {
+        Self {
+            username: username.into(),
+        }
+    }
+}
