+++ src/servers/Cargo.toml
@@ -14,6 +14,7 @@ aide = { version = "0.9", features = ["axum"] }
 api = { workspace = true }
 arrow-flight.workspace = true
 async-trait = "0.1"
+auth.workspace = true
 axum = { version = "0.6", features = ["headers"] }
 axum-macros = "0.3.8"
 base64 = "0.13"
